609	原来long long的运算速度比int慢一倍左右; change back;

+++ npbenchmark-main/SDK.PCP/pcp_vector.cpp
@@ -351,7 +351,7 @@ void PCP_Vector::find_pair()
                 // obj: num of vertices that has not been covered;
                 // Meaning: this move is better than history best;
                 // cerr << endl << "i " << i << " j " << j;
-                long long int this_iter_delta = center_weights[j] - center_weights[ic];
+                int this_iter_delta = center_weights[j] - center_weights[ic];
                 // cerr << " f(X+{i}-{j}) " << f_X_i_j << endl;
                 if(this_iter_delta < min_delta)
                 {
